requirements.txt
cloudscraper==1.2.58
lxml==4.6.4
PySimpleGUI==4.53.0
pystashlib==0.4.4
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
